requirements.txt
progressbar2
chainside-btcpy
pandas
ecdsa
